clinical: refactor barcode QC to shared module
Move the `barcode_quality_control()` function, helper functions, and log from
the `clinical` command to a shared module for future use by other
commands.
