+++ lib/seattleflu/db/cli/command/__init__.py
@@ -1,6 +1,7 @@
 """
 Commands for the database CLI.
 """
+import logging
 import pandas as pd
 
 
@@ -13,9 +14,63 @@ __all__ = [
 ]
 
 
+LOG = logging.getLogger(__name__)
+
+
 def add_metadata(df: pd.DataFrame, filename: str) -> pd.DataFrame:
     """ Adds a metadata column to a given DataFrame *df* for reporting """
     df['_metadata'] = list(map(lambda index: {
         'filename': filename,
         'row': index + 2}, df.index))
     return df
+
+
+def barcode_quality_control(clinical_records: pd.DataFrame, output: str) -> None:
+    """ Perform quality control on barcodes """
+    missing_barcodes = missing_barcode(clinical_records)
+    duplicated_barcodes = duplicated_barcode(clinical_records)
+
+    print_problem_barcodes(pd.concat([missing_barcodes, duplicated_barcodes],
+                                 ignore_index=True), output)
+
+    assert len(duplicated_barcodes) == 0, "You have duplicated barcodes!"
+
+
+def missing_barcode(df: pd.DataFrame) -> pd.DataFrame:
+    """
+    Given a pandas DataFrame *df*, returns a DataFrame with missing barcodes and
+    a description of the problem.
+    """
+    missing_barcodes = df.loc[df['barcode'].isnull()].copy()
+    missing_barcodes['problem'] = 'Missing barcode'
+
+    return missing_barcodes
+
+
+def duplicated_barcode(df: pd.DataFrame) -> pd.DataFrame:
+    """
+    Given a pandas DataFrame *df*, returns a DataFrame with duplicated barcodes
+    and a description of the problem.
+    """
+    duplicates = pd.DataFrame(df.barcode.value_counts())
+    duplicates = duplicates[duplicates['barcode'] > 1]
+    duplicates = pd.Series(duplicates.index)
+
+    duplicated_barcodes = df[df['barcode'].isin(duplicates)].copy()
+    duplicated_barcodes['problem'] = 'Barcode is not unique'
+
+    return duplicated_barcodes
+
+
+def print_problem_barcodes(problem_barcodes: pd.DataFrame, output: str):
+    """
+    Given a pandas DataFrame of *problem_barcodes*, writes the data to
+    the log unless a filename *output* is given.
+    """
+    if output:
+        problem_barcodes.to_csv(output, index=False)
+    else:
+        problem_barcodes.apply(lambda x: LOG.warning(
+            f"{x['problem']} in row {x['_metadata']['row']} of file "
+            f"{x['_metadata']['filename']}, barcode {x['barcode']}"
+        ), axis=1)
